+++ internal/pkg/sql/parser/factory/factory.go
@@ -3,8 +3,8 @@ package factory
 import (
 	"fmt"
 
-	"github.com/patterninc/heimdall/pkg/sql/parser"
-	"github.com/patterninc/heimdall/pkg/sql/parser/trino"
+	"github.com/patterninc/heimdall/internal/pkg/sql/parser"
+	"github.com/patterninc/heimdall/internal/pkg/sql/parser/trino"
 )
 
 type ParserType string
