+++ internal/pkg/sql/parser/trino/TrinoLexer.g4
