+++ internal/pkg/sql/parser/trino/parser.go
@@ -4,8 +4,8 @@ import (
 	"log"
 
 	"github.com/antlr4-go/antlr/v4"
-	"github.com/patterninc/heimdall/pkg/sql/parser"
-	"github.com/patterninc/heimdall/pkg/sql/parser/trino/grammar"
+	"github.com/patterninc/heimdall/internal/pkg/sql/parser"
+	"github.com/patterninc/heimdall/internal/pkg/sql/parser/trino/grammar"
 )
 
 type TrinoAccessReceiver struct {
