+++ internal/pkg/sql/parser/trino/tests/delete_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/patterninc/heimdall/pkg/sql/parser"
-	"github.com/patterninc/heimdall/pkg/sql/parser/trino"
+	"github.com/patterninc/heimdall/internal/pkg/sql/parser/trino"
 )
 
 func TestParseSQLDelete(t *testing.T) {
